New Handlers struct to house handlers for Chi
Put handlers under a struct to maintain a shared state, it's
currently a wireguard interface. Without help of a struct, we need
to re-intialize the interface in every handlers, or maintain a global var.
Both ways are not good. Moreover, we'll depend on not-so-related config
package to have information to initialize the interface. This in turn
raises another issue: config package shouldn't used directly, it need
to be loaded with a flag to a config file to load.

With help of a struct, we now can inject what's needed to struct's
New function