Stop abusing pointers everywhere
I was wrong about using pointers. I thought it would improve
performance by not coping values over to func's environment.
This is immature-optimization because we're not facing performance
issue yet. However, this pointer-abusing causes a problem: we dont
know if a func changes a pointer's reference value or not, so we'll
lose confidence when using the func.

The bottom line is, only passing pointers as func param only
when necessary, e.g we need to update the value of such param.